:recycle: Refactor(Device Middlewares): Refactored device POST ... Refactoring validations of tags field and version field from route POST a new device.
